+++ config/int_test.go
@@ -1,0 +1,54 @@
+package config_test
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/dogmatiq/dodeca/config"
+	. "github.com/dogmatiq/dodeca/config"
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+var _ = Describe("func GetInt64()", func() {
+	It("returns the integer value", func() {
+		b := Map{"<key>": String("123")}
+
+		v, ok, err := GetInt64(b, "<key>")
+		Expect(err).ShouldNot(HaveOccurred())
+		Expect(v).To(BeEquivalentTo(123))
+		Expect(ok).To(BeTrue())
+	})
+
+	It("sets ok to false if the key is not defined", func() {
+		b := Map{}
+
+		_, ok, err := GetInt64(b, "<key>")
+		Expect(err).ShouldNot(HaveOccurred())
+		Expect(ok).To(BeFalse())
+	})
+
+	It("returns an error if the value can not be parsed", func() {
+		b := Map{"<key>": String("<invalid>")}
+
+		_, _, err := GetInt64(b, "<key>")
+		Expect(err).To(MatchError(`<key> is not a valid signed 64-bit integer: strconv.ParseInt: parsing "<invalid>": invalid syntax`))
+	})
+})
+
+func ExampleGetInt64() {
+	os.Setenv("FOO", "123")
+
+	v, ok, err := config.GetInt64(config.Environment(), "FOO")
+	if err != nil {
+		panic(err)
+	}
+
+	if !ok {
+		fmt.Println("key is not defined!")
+	} else {
+		fmt.Printf("the value is %d!\n", v)
+	}
+
+	// Output: the value is 123!
+}