+++ CHANGELOG.md
@@ -14,8 +14,9 @@ The format is based on [Keep a Changelog], and this project adheres to
 ### Added
 
 - Add `Map`, an in-memory implementation of `Bucket`
-- Add `GetBool()` and its variants for reading `bool` configuration values
-- Add `GetInt64()` and its variants for reading `int64` configuration values
+- Add `GetBool()` and its variants for reading boolean configuration values
+- Add `GetInt()` and its variants for reading integer configuration values
+- Add `GetFloat[32|64]()` and their variants for reading floating-point configuration values
 
 ## [0.2.2] - 2020-09-04