+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [1.0.0-rc.0] - 2020-09-10
 
 ### Added
 
@@ -67,6 +67,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.2.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.2.0
 [0.2.1]: https://github.com/dogmatiq/dogma/releases/tag/v0.2.1
 [0.2.2]: https://github.com/dogmatiq/dogma/releases/tag/v0.2.2
+[1.0.0-rc.0]: https://github.com/dogmatiq/dogma/releases/tag/v1.0.0-rc.0
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD